Sargis/pid

+++ src/firmware/control/attitude_adjust/attitude_adjust.h
@@ -16,6 +16,6 @@
 uint8_t InitializeAttitudeAdjust(void);
 void AttituteAdjustSetDesired(double yawDesired, double pitchDesired, double rollDesired);
 void AttituteAdjustUpdatePID(double yawActual, double pitchActual, double rollActual);
-void AttitudeAdjustGetError(int16_t motor_changes[NUM_MOTORS]);
-void AttitudeAdjustSetActuation(int16_t motor_changes[NUM_MOTORS]);
+void AttitudeAdjustGetError(int motor_changes[NUM_MOTORS]);
+void AttitudeAdjustSetActuation(int motor_changes[NUM_MOTORS]);
 #endif
